Moving components into dynamicly loaded components to help with load times. I also updated a colour on the new MPDX as it was slightly off from the MPDX logo
